Clean up headers and deps for visualization and replica checks
