+++ builder/test/BUILD
@@ -6,7 +6,10 @@ package(
 cc_library(
     name = "util",
     testonly = True,
-    srcs = ["context.cc"],
+    srcs = [
+        "context.cc", 
+        "util.cc",
+    ],
     hdrs = [
         "context.h",
         "funcs.h",
@@ -14,10 +17,11 @@ cc_library(
     ],
     deps = [
         "//runtime",
+        "//runtime:visualize",
         "//base:chrome_trace",
         "//base:thread_pool",
         "//base/test:util",
-        "@bazel_tools//tools/cpp/runfiles",
+        "//builder:replica_pipeline",
         "@googletest//:gtest",
     ],
 )
@@ -65,10 +69,7 @@ cc_test(
         ":util",
         "//base/test:util",
         "//builder",
-        "//builder:replica_pipeline",
         "//runtime",
-        "//runtime:visualize",
-        "@bazel_tools//tools/cpp/runfiles",  # Needed for -DBAZEL_CURRENT_REPOSITORY even though it's used transitively via util
         "@googletest//:gtest_main",
     ],
     size = "small",
@@ -84,7 +85,6 @@ cc_test(
         "//builder",
         "//builder:replica_pipeline",
         "//runtime",
-        "@bazel_tools//tools/cpp/runfiles",  # Needed for -DBAZEL_CURRENT_REPOSITORY even though it's used transitively via util
         "@googletest//:gtest_main",
     ],
     size = "small",
@@ -100,7 +100,6 @@ cc_test(
         "//builder",
         "//builder:replica_pipeline",
         "//runtime",
-        "@bazel_tools//tools/cpp/runfiles",  # Needed for -DBAZEL_CURRENT_REPOSITORY even though it's used transitively via util
         "@googletest//:gtest_main",
     ],
     size = "small",
@@ -109,6 +108,7 @@ cc_test(
 cc_test(
     name = "softmax",
     srcs = ["softmax.cc"],
+    data = glob(["visualize/*.html"]),
     deps = [
         ":util",
         "//base/test:util",
