+++ builder/test/softmax.cc
@@ -1,7 +1,6 @@
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
 
-#include "base/test/bazel_util.h"
 #include "builder/pipeline.h"
 #include "builder/test/funcs.h"
 #include "builder/test/util.h"
@@ -10,25 +9,6 @@
 #include "runtime/visualize.h"
 
 namespace slinky {
-
-void check_visualize(const std::string& filename, const pipeline& p, pipeline::buffers inputs,
-    pipeline::buffers outputs, const node_context* ctx) {
-  std::stringstream viz_stream;
-  visualize(viz_stream, p, inputs, outputs, ctx);
-  std::string viz = viz_stream.str();
-
-  std::string golden_path = get_bazel_file_path("builder/test/visualize/" + filename);
-  if (is_bazel_test()) {
-    std::string golden = read_entire_file(golden_path);
-    ASSERT_FALSE(golden.empty());
-    // If this check fails, and you believe the changes to the visualization is correct, run this
-    // test outside of bazel from the root of the repo to update the golden files.
-    ASSERT_TRUE(golden == viz);
-  } else {
-    std::ofstream file(golden_path);
-    file << viz;
-  }
-}
 
 // This implementation of softmax is not intended to be fast, it is only intended to model the data dependencies.
 index_t max_dim0(const buffer<const float>& in, const buffer<float>& max_in) {
